audit: fix 11 issues found by full-project ultracode review
A workflow audit (8 parallel scanners → adversarial verify, run wf_460cd277-839)
flagged 134 raw findings, of which 27 critical/high were confirmed real. This
commit applies every mechanical fix; the 7 remaining real findings need
design judgment and are tracked separately.

CRITICAL — frontend crash
  frontend/src/pages/DashboardPage.jsx: added the missing
  `import { ResponsiveContainer, LineChart, CartesianGrid, XAxis, YAxis,
  Tooltip, Line } from 'recharts'`. Lines 571-582 used those identifiers bare;
  any user with ≥2 trend points was crashing the dashboard with a
  ReferenceError.

HIGH — deploy regression risk
  deploy-backend.ps1: bumped --memory from 1Gi to 4Gi. Production was bumped
  to 4Gi after the 2026-06-24 OOM; running the deploy script would have
  silently rolled the next revision back to 1Gi and OOM'd on first RAG load.

HIGH — silent advice truncation
  backend/api/services/ai_engine.py:308: changed
  `re.sub(r'\[\s*參考\s*\d+\s*\][^\n]*', '', reply)` →
  `re.sub(r'\[\s*參考\s*\d+\s*\]', '', reply)`. The `[^\n]*` tail was
  greedily eating every character after a `[參考N]` citation through
  end-of-line, silently deleting the model's real advice. Comment claimed it
  only stripped the marker.

HIGH — SSRF via push subscription
  backend/api/views/dashboard.py: added _is_allowed_push_endpoint() allowlist
  (FCM / Mozilla / Windows Notification / Apple Push). Previously any
  authenticated user could register an internal / metadata URL and pivot
  pywebpush into outbound HTTP calls from inside the trust boundary
  (notifications fire on friend request, achievement, weekly summary).

HIGH — authenticated email enumeration
  backend/api/serializers.py: dropped `email` from UserSearchSerializer
  fields/read_only_fields. Any logged-in user could harvest emails by
  iterating substring searches (PII / PDPA / GDPR). Frontend friends.js had
  zero references to the field.

MEDIUM
  * backend/api/views/health.py:455 — avg_deep_pct denominator and
    numerator now come from the same filtered record set, so the
    user-facing insight string can no longer exceed 100%.
  * backend/api/tests.py:366 — added ThrottleTests.setUp() with
    cache.clear() so DRF throttle counters don't bleed between test
    classes (matches the AchievementTests pattern).
  * backend/api/models.py:543 + new migration 0057 — widened
    AuditLog.target_id from IntegerField to BigIntegerField. All PKs are
    BigAutoField; a signed 32-bit target_id would silently overflow once
    a referenced PK exceeded 2.1B.
  * Makefile — added `make test-clean` target that runs the full suite
    WITHOUT --keepdb (CI parity; catches migration drift locally).
  * .env.example — documented 13 env vars that settings.py / views read
    but were absent from the example: GOOGLE_CLIENT_ID/SECRET,
    CRON_SECRET, SENTRY_DSN, VAPID_PRIVATE_KEY, GS_BUCKET_NAME,
    CSP_API_HOST/WS_HOST, EMAIL_USE_SSL, THROTTLE_LOGIN_PER_USERNAME,
    CELERY_BROKER_URL, WS_ALLOW_QUERY_STRING_TOKEN,
    CHROMA_DISABLE_AUTO_BOOTSTRAP, DISABLE_AI_PREWARM,
    HEARTBOX_ML_MODEL_DIR.

LOW
  * frontend/src/components/FeedbackWidget.jsx — guarded the 2s
    post-submit setTimeout with a ref + useEffect cleanup so navigating
    away doesn't fire setState on an unmounted component.

VERIFICATION
  * 259/259 Django tests pass under test_settings (no --keepdb)
  * Frontend `npm run build` succeeds with the new recharts import
  * Migration 0057 generated by Django's makemigrations (not hand-written)
  * No semantic changes to existing code paths beyond the surgical edits

DEFERRED — real but need design judgment, escalated to user (see workflow
log wf_460cd277-839): DailySleepView serializer bypass, test_apps_prewarm
tautological assertion, ThrottleTests disjunctive assertions, missing
crisis_guard E2E test, NotificationWebSocketPushTests misnamed, migration
0056 robustness, migration 0009 silent decrypt failure repair.

+++ backend/api/models.py
@@ -540,7 +540,10 @@ class AuditLog(models.Model):
     )
     action = models.CharField(max_length=30, choices=ACTION_CHOICES)
     target_type = models.CharField(max_length=50, blank=True, default='')
-    target_id = models.IntegerField(null=True, blank=True)
+    # BigInt because every PK in this schema is BigAutoField (8-byte); a
+    # signed 32-bit target_id would silently overflow once any referenced
+    # row's PK exceeds 2.1B.
+    target_id = models.BigIntegerField(null=True, blank=True)
     ip_address = models.GenericIPAddressField(null=True, blank=True)
     details = models.JSONField(default=dict, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
